+++ CHANGELOG.md
@@ -36,6 +36,8 @@
 - Add a ProductQueryFactoryInterface argument in ProductDatasource::__construct
 - Add a $productOrmAdapterClass argument in DatasourceAdapterResolver::__construct
 - Remove is_default, translatable from attributeOption mapping
+- Remove AttributeOption::setDefault, AttributeOption::isDefault
+- Remove AttributeInterface::getDefaultOptions
 - Remove $optionClass and $optionValueClass arguments from the AttributeManager::__construct
 - Remove createAttributeOption, createAttributeOptionValue, getAttributeOptionClass from the attributeManager (now in the attributeOptionManager)
 - Add a $attributeOptionManager argument in AttributeController::__construct